feat: add searchState param to more queries (#183)
#### What is the purpose of this pull request?

To be able to use the `searchState` param in more queries. This enables
us to send user data into the search-resolver app easily from the
front-end. Added to the following queries in this PR:
- `searchSuggestions`
- `products`
-  `productRecommendations`
- `topSearches`
- `autocompleteSearchSuggestions`

#### What problem is this solving?

Currently, there is no way to get data from the front-end regarding the
user's session into the search resolver directly through these queries.
Having this feature to get data this way would help in analytics and to
provide better and customized results to the user.

#### How should this be manually tested?

[Workspace](https://usercookie--worldwidegolf.myvtex.com/)

#### Types of changes

* [ ] Bug fix (a non-breaking change which fixes an issue)
* [x] New feature (a non-breaking change which adds functionality)
* [ ] Breaking change (fix or feature that would cause existing
functionality to change)
* [ ] Requires change to documentation, which has been updated
accordingly.