feat: add options data to productSearch query response (#191)
#### What is the purpose of this pull request?

This PR adds `options` data to productSearch query response.
[Task TIS-59](https://vtex-dev.atlassian.net/browse/TIS-59)

#### What problem is this solving?

We need to use the deliveryPromisesEnabled data to enable the delivery
filter in the PLP. This is necessary to make it easier to activate
shipping filters in a store.

#### Screenshots or example usage

<img width="416" alt="image"
src="https://github.com/user-attachments/assets/d16b7042-7086-474c-a282-1e0905e9c67d"
/>


#### Types of changes

* [ ] Bug fix (a non-breaking change which fixes an issue)
* [x] New feature (a non-breaking change which adds functionality)
* [ ] Breaking change (fix or feature that would cause existing
functionality to change)
* [ ] Requires change to documentation, which has been updated
accordingly.