Feat/tis 302 (#195)
This pull request updates the GraphQL queries for product search and
suggestions to support an additional `origin` parameter and exposes a
new `searchId` field in the product suggestions response. These changes
enable more flexible querying and tracking of search requests.

**GraphQL query enhancements:**

* Added an optional `origin` parameter to the `productSearchV3` query
definition and passed it to the server, allowing clients to specify the
source of the search request.
[[1]](diffhunk://#diff-8a40d928c4aa83928add5b55bbe2621b0d679b6cb6cab0ea7bd19d16d23ddfe7R27)
[[2]](diffhunk://#diff-8a40d928c4aa83928add5b55bbe2621b0d679b6cb6cab0ea7bd19d16d23ddfe7R45)
* Added an optional `origin` parameter to the `productSuggestions` query
definition and included it in the server request, enabling source
tracking for product suggestions.
[[1]](diffhunk://#diff-9c5fa35f47378068fdf019af8b0ecfb352daf50b5fbb4c21f341aeb25fa6a7e6R11)
[[2]](diffhunk://#diff-9c5fa35f47378068fdf019af8b0ecfb352daf50b5fbb4c21f341aeb25fa6a7e6R25)

**Response data improvements:**

* Exposed the `searchId` field in the `productSuggestions` query
response, allowing clients to uniquely identify search sessions.